backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic>=2.7.4,<3
pydantic-settings==2.1.0
python-dotenv==1.0.0
langchain==0.3.13
langgraph==0.2.60
langchain-google-genai==2.0.8
langchain-openai==0.2.14
python-multipart==0.0.6
aiofiles==23.2.1
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
